+++ lib/express/network/http2/client.ex
@@ -22,6 +22,11 @@ defmodule Express.Network.HTTP2.Client do
   @callback send_request(pid(), list(), String.t) :: {:ok, pid()} | any()
 
   @doc """
+  Sends a ping through a socket by `pid`.
+  """
+  @callback ping(pid()) :: :ok
+
+  @doc """
   Receives a response from socket by `pid` for a `stream`.
   """
   @callback get_response(pid(), pid()) :: {:ok, {String.t, String.t}} | any()